backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
aiosqlite==0.20.0
greenlet==3.0.3
pydantic==2.7.1
python-dotenv==1.0.1
websockets==12.0
httpx==0.27.0
python-jose[cryptography]==3.3.0
bcrypt==4.1.3
firebase-admin==6.5.0
azure-communication-sms==1.0.1
slowapi==0.1.9
structlog==24.1.0
azure-communication-identity==1.5.0
azure-communication-call-automation==1.3.0
